fixed minor bug that did not let player power smash crates when speed was <0

+++ Assets/Board/Scripts/Cell.cs
@@ -51,7 +51,7 @@ public class Cell
         this.type = type;
     }
 
-    public void smashCrate(int speed)
+    public void smashCrate(float speed)
     {
         if (type != CameraScript.Element.CRATE) return;
         if (overlayObj == null) return;
